Put stuff back in attrs but hide the things we don't want
